refactor(rebalancer): move utils up in the tree
